Add support for Proxmox Backup Server storage backend

+++ lib/backup/cleaner.rb
@@ -49,6 +49,8 @@ module Backup
       ##
       # Remove the temporary folder used during packaging
       def remove_packaging(model)
+        return if model.storages.length == 1 && model.storages.first.is_a?(Backup::Storage::ProxmoxBackupServer)
+
         Logger.info "Cleaning up the temporary files..."
         FileUtils.rm_rf(File.join(Config.tmp_path, model.trigger))
       end
